And most of the work for inbox; for a toy one, at least.
I'll still need to add some things. And clean things up; it feels
unnecessarily hard to parse, currently. I can't shake the feeling
there's a less verbose way to put all of this.

In any case, probably do the INBOX table, next, and then write to it,
on success. Then I can try having something follow the user and,
hopefully, it'll show up in the INBOX and I can take a look at it.