The `INBOXES` table. This'll probably undergo some alteration, as time goes. It's not my job to keep track of other people's posts (and, if performant, ideally just look up the original object for info. on it; in theory, this'll be able to be used to allow editing of objects…) so we're just going to store the JSON directly under the `ACTIVITY` column.
I included the types as their own column since I could see wanting to
filter out, say, only likes. Thinking about this now, though, there is
a `likes` collection….

It may be better, anyway, not to duplicate the data, like that, to
another table and it may, still, be useful to filter on particular
activities (like getting all creates or something like that…). We'll
see. I hate having to double back and change things but that can't be
helped, with programming. Things may alter, in the future.