Frontend will need to let backend processes know which user is making the request. So we'll save, as a cookie, the username. I was worried about someone editing it and then trying to, like, have people cause other people to publish posts or something but I realized that I could check if the cookie is accurate by checking if a session with the username actually exists.
I have to maintain two sessions (one with a general name so it's
searchable in any context and one…; hmm. I suppose, if we considered
the username cookie to be present at all circumstances that a user is
logged in, we could just pull up the cookie, check if a matching
session name exists and, if so, we consider the cookie and session to
be valid. I'll have to change this but I kinda like that; I'll have to
check whether Artanis is considering sessions per browser, etc. rather
than for anything hitting the service so I can make sure someone
doesn't just have to change the cookie (though I'll try to make
editing cookies as difficult as possible…) to a username of someone
logged-in in a totally different country to get into their account;
but, if so, then I rather like this) and nevermind.