Until I can get these redirects under control, it's safer to just redeploy with the user I made, this time.
I'm starting to wonder if the session stuff is off, considering I
think I ran into weird behavior like this last time and it doesn't
seem to be able to recognize when I user's signed in. Could be my
code, though; I haven't exactly taken a close look at the old stuff
when moving it to this branch.