I need to see if I can do the same for non-`GET` calls so things can be uniform.
I'm not sure it really adds anything other than asthetic of the code;
but, then again, that was/is a huge draw of Ruby on Rails.

I dunno; I think the real benefits (or lack thereof) will be more
evident once I can get the `post` calls switched over, as well.