I'd like, if I can, to not use as much JavaScript as possible for this site; _but_ (naturally) a lot of the different pages'll have the same layout since the idea is to work from out of the main page.
So I'll need to inject that as a snippet (rather than rewrite it a
million times) and I think it makes more sense to call that a
dashboard; perhaps…. Either way, every other page that shares the side
buttons, side bar, etc. can lay claim to the name "dashboard" so it's
better to define this as what function it's serving (it's the main
function: viewing the timeline).