Some colors that pass AAA spec.s.
Next to the default gray text, they don't really stand out so they'll
probably have to be used for icons and the like.

I dunno what I'll do for highlights in the context of text and the
like.

The `input` already uses a highlight that's not even AA-matching
against the white to stick out enough against the gray.

We'll see; I can get feedback from others and guidance about the best
way to go.