It's recommended by the AP spec. (and elsewhere) to store Actors (or perhaps it was Objects, in general; Actors are Objects and there's enough similarity in the concept, even if they weren't, so we'd do this, either way) and it wouldn't make sense to not store local Actors with the remote Actors, given they're the same type.
So let's adjust the databases, a bit.

Move all attributes that aren't unique to local Actors into the ACTORS
table (sans the username since it's, technically, used as a unique
key; I forget why I didn't just make it the primary key…. It doesn't
have to be a `text` type so I could index on it…).

And add a column to link to the ACTORS table.

Also, add the ACTORS table; I think I'll need a ENDPOINTS table and
then that should cover everything required by the spec. We'll probably
also add stuff for attachments, images, etc. like Masto has; but we
can do that at a later date.