While I was thinking over whether to store a Actor's public key in a separate table or not, I realized I had already concluded that it didn't matter storing it as-is because it's, well, already public. So I've decided to reverse encrypting the public key in the database.
I'm going to opt for not storing the public key in the database, for
now, since I don't quite grasp the full ramifications of such a
decision, yet….

I'm still not sure how to handle some aspects of storing remote Actors
in the database; I've seen several places say to do it so it seems
like the right choice to make.

It's just that, unlike Objects like Notes and Articles, I don't
_think_ there are any mechanisms in place to notify when a change
occurs (like deleting or editing like with the aforementioned
Objects).

Perhaps I just don't quite understand how to use this setup, yet, and
it'll become more clear to me later (or I misread the things that I
had and they only meant to store Objects, not Actors, in the database
and I'll have to rewrite all of this, again; fun, fun!).