I think this makes sense; CCs and TOs will be stored in their own mapping tables, similar to FOLLOWING and FOLLOWERS. Just a simple Object to Actor mapping (since any Activity should have the same addressing as their object, I believe; we may end up having to edit this).
The one snag I can think of is that I wasn't planning on storing
Actors in my Objects table.

But maybe that was a dumb idea, to begin with…; like, sure, Note and
Article are probably similar but more dissimilar Objects may not be
able to properly encapsulate everything in one table….

Should I store all the Notes in a NOTES table, Articles in an ARTICLES
table, etc.? It wouldn't make for easy reverse lookup… (everything
referencing an Object would have to reference the OBJECTS table but
there's no easy way to then reference specific tables; like, I'd have
to have backend logic that gets updated every time we add a new Object
that we support that'd have to look at the OBJECT_TYPE and determine
which table to look it up in and then a separate column with numerical
IDs; not to mention that there's no way to slap a foreign key on
that (not even from the specific table to the OBJECTS table; since
there's multiple specific tables, the IDs could be (and probably will
be) non-unique); gross…. Actually, I could store the OBJECT_ID in the
specific table and, say, – upon finding it's a Note – search for the
OBJECT_ID in the NOTES table; but I still don't like the manualness of
that…).

We'll have to see; _hopefully_, the possible attributes of a Object
are all the same and the type just signifies that you should figure
out how to display it; then I could just store everything in the one….