Since an Actor is an Object and an User is just an internal Actor, we can simplify things by just having having Actors and Users use the OBJECT_ID for their primary key.
I'd, originally, been considering calling the column out by its actual
table value (OBJECT_ID) to remind developers of where this value is
coming from but I realized, especially when operating in the context
of the OBJECTS table, that it might get confusing which table we're
actually referring to referencing OBJECT_ID.

While often it'll be in the context of grabbing the ID from the
respective table (so something like `(assoc-ref user "OBJECT_ID")`), I
think it'll be clearer overall to just refer to the ID of the ACTORS
table as ACTOR_ID; it's easier to remember just one rule of "USER_ID
and ACTOR_ID are OBJECT_ID". Once the ability of setting some sort of
foreign key implementation is in place for Artanis, we can better make
that connection more clear, too.