This should cover all the available attributes for Objects (with the intention to properly implement these at a later date).
I think I'm going to keep the Actors in the `OBJECTS` table.

Since there's so much overlap, we can cover all the inherited Object
attributes in `OBJECTS` and put Actor-specific attributes in the
`ACTORS` table (we can probably remove the `AP_ID` and `ACTOR_TYPE` to
accomplish that fully but we'll do that once we're able to do proper
joins as SQL calls with Artanis).

This way, also, doesn't change as much of the code that's already written.