I noticed a lot of the same code was happening repeatedly (and it so happened to be the largest block, too).
So we'll move this to its own file; hopefully the start of organizing
database operations into something more abstract…

We'll avoid looking up the `attributedTo` property for Actors; I
think, technically, they're allowed to have them so I'll have to do it
eventually, I'm sure, but not right now.

There's also the fact that it isn't necessarily an Actor; so I'm
making a lot of assumptions, already.

This might have to be reorganized.

But this'll work to jump-start stuff, right now.