And something similar, for setting data.
The thing that's particular nice about this one is we can centralize
things like reversing the ActivityPub ID or checking whether the
optional attributes are there (and, if not, provide `'null`) or
replacing single quotes in strings with two single quotes.

Then I don't have to worry about checking any of that crap and can
just concern myself with treating the data straight-forwardly as it's
supposed to be thought of, knowing that the functions will take care
of any cleanup for me.