And add these in.
Also, add in the JavaScript to show and hide.

I was _so_ hoping to be able to not have to use JavaScript but the CSS
wouldn't seem to behave.

I tried with a hidden checkbox and label but, even though (I think) I
had the CSS to target all succeeding siblings (`+`), it only ever
affected the most immediate siblings.

And we had to hide the footer and show the text area. And I didn't
want to populate the thing with a ton of extra tags.

I dunno; probably for the best, seeing as we should probably populate
the `textarea` with the users you're replying to and I don't see any
way to do that one without JavaScript…