'Had to put this in here, rather than the getters, due to `insert`ing stuff (much to my displeasure…) but this'll be useful and, hopefully, both centralize and simplify outside logic.
If we can find the ActivityPub ID (or IDs) in our database, simply
return.

But, if not, pull down the info. and feed it to `insert-actor-auto` so
we can add it and get back our database ID.

Like the adjustment I made to `get-object-dbID-by-apID`, it'll give
back a single result if you give it a single ActivityPub ID and give
back a list if you give multiple.

I dunno how I feel about a variable response, like this since it could
get confusing when all the other getters always return a list but
maybe the consistency of naming should be enough to remind…

I can always adjust the function name, too, to better capture the behavior.