I realized (belatedly…) that I'm never going to pass a string to this (I should never).
All other parameters of this function expect to receive an Scheme
object in its final form so we should – here –, as well.

Also, let's handle if the `actor` attribute is a single actor, instead
of list, here, at the root of activity-handling stuff, so we don't
have to worry about it anywhere else.