Time to simplify this.
We don't need the Activity type or Actor/Object IDs – those are all
stored with the Activity. We just need to designate which user('s inbox) the
particular activity belongs to.

So I guess it's more of a mapping table; but that's O. K.

This should, really, be all we need.