Weird…crap is going on.
Seems the issue is on line 276 of
`database-setters.scm` (specifically, `get-actor-dbID-by-apID`).

The issue is an error I should have caught (and part of why I was
worried about mixing returning lists and singular values for
`get-actor-dbID-by-apID` and `get-object-dbID-by-apID`…).

`get-actors-where` will return an empty list, if it can't find
anything in the database; so I need to check for `null?`.

But I'm seeing something weird in the logs about what it's checking
for.

It's a truncated ActivityPub ID with a question mark, at the end.

Maybe it's just the log looking something weird; but I want to double
check that it's not actually trying to look up this janky AP ID.

So let's deploy this and check, just to be sure. Then fix the bug.