Welp; this explains why things were acting weird…
It helps if you submit the right parameters in the right order to the
function you're trying to call.

I don't think the `actor` of the activity should be added, yet, so I'm
going to keep the bug, for now, just to test what the data looks like
before I go fixing things.