That looks to've worked. I'm getting actual proper `actor`s, now.
But we also seem to have 29 Objects in our database, after, like,
three messages…

(1) There's the temporary account I did to test signing up
(2) The real user account
(3) The activity of the first message
(4) The object of the first message (which I don't think we even got
to, yet
(5) The activity of the second message
(6) The object of the second message
(7) The activity of the third message
(8) The object of the third message

Aaaand…that should be it? If things are failing at the `insert-actor`
part, it should have only reached adding the object _of_ the activity
and not managed to even add the Actor, yet.

I guess, if the requesting server keeps trying again, I might end up
with more.

But, still, that's likely 5 objects, on first try, and then (let's
assume they try 3 times) 11 objects, in total.

I dunno; I'll have to check out the database, I think…