GOSH FUCKING DAMMIT!
I wonder how much of this is me having to write everything for the
first time for a project, me being used to working in a (very
functional…) object-oriented programming paradigm most of the time, or
just something that'd be easier to avoid if I _were_ using a
strongly-typed and/or object-oriented language.

I dunno; I could also just be writing sloppy code…

These may be things I could fix with enough tests…it's not like they
should be _changing_ any, at any point.