My…less than pleased implementation of storing images.
So, because an object can have more than one of an image or an icon, I
can't just (cleanly) link to an `OBJECT_ID`; we need a mapping table,
of some sort…

So, same premise as usual – an Image is (a Document which is) an
Object so the `IMAGE_ID` links to an `OBJECT_ID` in `OBJECTS` which
stores the Object fields.

The examples given in the ActivityStreams documentation show icons
having `width`s and `height`s, even though the documentation
explicitly only provides such attributes to Links (le sigh…) so, while
we have to have a separate table (anyway), provide the ability to
store such data.

And a final `IS_ICON` column so I don't have to create two tables to
store identical objects while still being able to distinguish between
which attribute of an object they belong to.

Primary key on the `IMAGE_ID` and `OBJECT_ID` since, theoretically,
someone could use the same image for both an `icon` and `image` of two
different Objects; rather than duplicating every URL we get of an
Image, we can store the URL, media type, etc. once in the `OBJECTS`
table and allow duplication in `IMAGES` to capture the fact that the
same Object is, say, an `icon` for one Object and an `image` for
another.

Fun stuff.