And the bulk of this update.
I'm getting more uneasy about this repetition of code that's occurring.

I know I originally ventured down seeing how this would work since,
while an Actor – say – is built on an Object, I could see certain
entities treating certain properties differently than their
supertypes.

But that hasn't really been the case, that I've seen. I'm winding up
with more and more code that's repetitive…

And update to one attribute's going to mean a crap-ton of update
everywhere for code that's effectively the same thing.

But I don't see an easy way to cleanly merge it all, either…

In any case, this allows to search for particular Images.

Since the `IMAGES` database has an `OBJECT_ID` column to link the
image back to a parent Object, we assume that both searches for
`IMAGE_ID` and `OBJECT_ID` is from the `IMAGES` table
perspective (along with, of course, `WIDTH` and `HEIGHT`).

All other columns assume searching the `OBJECTS` table.

There may be other ways to handle this but the cleanest to write
seemed storing the `IS_ICON` column in the Image record and then being
able to filter on that in the final assessment (since we can't search
for whether it's such if we have to query `OBJECTS`, first; this is
something that should be able to be cleaned up once we find out how to
do `JOIN`s…).