Rather than leaving this up to whomever is sending things over, we can handle it in one location (I think we did a similar update, earlier, so this is in line with that logic).
I think it may be a good rule, going forward, to handle this stuff the
same way AP does – it doesn't clean up the data before passing it
over, it just expects us to handle whether it's a list or single
item. So let's do the same at the most root location for every time.