It dawned on me that, if we're allowing for Objects to not have AP IDs and we're still expecting we can check the URL to determine uniqueness, this function will not be able to handle, say, Images without any AP ID.
So we adjust this for it to take a look at the object and check if
there's any AP ID or `url`.

It prefers the AP ID, naturally, but, if one doesn't exist, uses the
`url` as the ID to find the object.

This does, of course, open up the possibility of someone using the
`url` for a different purpose than, say, an Image does but, if that's
the case, that's bad form on their part since (I assume) the reason
for not including an ID on Image, in the case of – say – Mastoton, is
that you can't look up the ID to get further info., like other
objects.

The image is the object and you should only care about the `url`.

If you're sending out objects with no ID but the same `url`, there's
the possibility of doing an update but, otherwise, it just doesn't
serve much purpose. Maybe you're not expecting them to be saved (in
which case, no harm here since their transitory, anyway).

We'll see; I may end up changing this, later.

The other benefit of this change is that it's able to handle a variety
of objects so I don't have to rely on fishing out the ID beforehand, anymore.