Forgot about my check at the end of `get-object-dbID-by-apID`…; it'll return a list of IDs for an alist.
Which doesn't seem to bother the SQL `INSERT`; but what happens if
that list has more than one value?

And, really, consistency of expectation for future developers is
important, as well.

Also, I think I really need to just decide whether we're dealing with
hash-tables or alists and stick to it; not any more of this waffling.

We can update the code if they decide to update the JSON library to
its most current edition.

This all might be easier to manage if I can clean up the code
repetition and the proliferation of the same code all over the place.