Macros to the rescue.
I would've liked to've have something where I could store the various
parameters in a variable (called `OBJECT_PROPERTIES` or something) so
then when I needed to update the parameters for an Object, I could
just update one spot and update the parameter for both `insert-entity`
and `insert-object` simultaneously.

I know I still have to update the body of `insert-object` for how to
handle each parameter but it would've meant 2 updates instead of 3, at
least.

Oh well; this is still, exceedingly, better.

Now to replace the other `insert-`s with this guy.