Took forever but here we go.
We can cut down the amount of repetition by just reusing the argument
list to generate both functions.

`insert-images` makes things a tad more complicated since it needs an
argument which isn't derived from the object, itself; so we have to
make sure the `afterRequiredArgs` include whether or not to include
each parameter in the equivalent `-auto` function…

But it's definitely better than writing everything out, every time.

With a centralized function – like this –, updating stuff should be trivial.