Using two separate macros kinda bothers me but I don't know of any way to combine them that would result in, either, reintroducing (albeit, less) repetition and 1,000 times more confusion/convolutedness.
Given that the merging together and fetching of the data from the
database wasn't particularly in a pattern, we'll leave those to be
manual.

There was, kinda, 3 different patterns that I noticed so, maybe (if we
can), we'll split it into three different macros, for each scenario?
But we'll see.