I should probably move this to the only file it's used but, in the meantime, I realized that this doesn't allow for finding elements that don't exist in the table.
For example, Images.

I won't find an "IMAGE" key in the data returned from the database.

Hence we pass `(const #t)`.

But that only handles the `check?`; the `assoc-ref` will still return
`#f`, resulting in this check failing.

Since Artanis does the right thing of returning a default value
when finding a `NULL` in the database, we probably can remove this
safely.

Also, we can just make sure to check for, say, 0, if we want to
exclude values that don't exist.