So, because `IMAGES` references the Object it belongs to with `OBJECT_ID`, this runs into the issue that we, typically, pull the DB ID by calling for `OBJECT_ID` (since all primary key ID columns share the same ID as their respective Object).
Even if we started calling by the particular table's primary key
column name, `OBJECTS` still uses `OBJECT_ID`.

So we need to remove it from the `IMAGES` data so we don't commit
ourselves to an infinite loop.

I'm O. K. with this since it's particular to Images (and not something
we have to do in every instance); so let's get this added.