Deploying out to Heroku and then trying to parse the errors generated there from the logs is just…too time consuming and not great for quick testing.
So let's add this to be able to test receiving Activities; skips the
whole checking-signatures part but that's O. K. (for now, at least).

I need to figure out a way to add tests in a way that doesn't create
infrastructure that's fragile and easy to break within a few changes.

I…guess…on to being able to send responses?

I feel like saving stuff to the Inbox still feels…half-formed. So it
feels weird to move onto handling the Outbox.

While I tampered back some of the unwieldiness, things still feel
unorganized and the Outbox, being something that's filled from a bunch
of different situations/places (unlike the Inbox where it just awaits
others to contact it and filters things out, from there; there's a
single point of entry).

It feels like one of those things that I'll be able to organize, as I
start programming for the parts, but that means unorganization, at the
beginning.

So it obviously makes sense to have the processes (like responding to
a post) put stuff in the Outbox, with adding Activities to an Outbox
being a separate function.

Maybe it makes more sense to design other elements of viewing a
site (like, being able to search for users and the like) first; that
won't require tackling the Outbox, remains "read"ing without any
"write"ing, and still adds more functionality that's sorely needed.

That'll strengthen a lot of that backend handling (such as, say,
communicating from the front-end to the backend and the like) and
further round things out; and, hopefully, organize things more along
the way so they're better by the time I tackle any Outbox-related
stuff.

Yeah; I might just do that…