Some work to handle favoriting posts.
It's…rather ugly.

Compounded by the fact that we can't use `http-post` and we need to
handle all signatures with `openssl` (via `system`).

I think, if those two were solved, a lot of this could be cleaner and
easier to read.

But oh well. For now, it gets the job done.

Basically, it'll make sure to save the `Like` Activity in the
`OBJECTS` and `ACTIVITIES` table (via `insert-activity-auto`) and then
update the `AP_ID`; I hate having to do it this way but, if we want
the ActivityPub ID to be based on the database, we need to…`INSERT`
it, first, to get an ID.

Then we just send it to the inbox of the user's post.

I'm sure we need to add some checks and the like, here. But this'll
do, for now. I just need to test it…