Again, severely enjoying the new macros we added.
This is a bit hairy (and in need to cleaning up) but the premise is
simple.

I realized that, while we have these posts we can display on the
timeline, we don't have an easy way to check if the user has liked
them or not.

I think creating a `post` record can make sense; there may be other
information we might want to bundle into a post when handling
rendering the HTML for them.

For now – though –, we'll just stick to storing a list of Users who've
liked the Object.

A lot of this can be cleaned up (code-wise and performance-wise) with
using `JOIN`s instead of querying the database individually (along
with `JOIN`s in the assembling of the User, as well; that should
settle things down to (roughly) two calls).

I don't know how much time creating the users (such as converting
strings to URIs, etc.) takes but I'd hope not long.

And it makes sense to provide the frontend with full user information,
in case they need it.

We'll see.