Not related to the problem, at hand, but my `gsub` function wasn't doing a replace-all, as I expected.
It still works, for the places it was originally used (since I need
regexp. for those locations) but it definitely wasn't cutting it for
weeding out single quotes in the JSON before committing them to the
database.

I knew there had to be a non-regexp. way to do this natively in
Guile. http://gnu-guile.7481.n7.nabble.com/Replacing-substrings-in-strings-td14202.html
lied to me…