That seems to've done the trick; it's no longer complaining about an invalid signature.
It _is_, however, now complaining about not being able to find the
public key.

This seems strange to me, seeing as the "publicKey" portion of the
JSON, here, is exactly identical to what Masto has their users (sans
the key itself, of course).

I'm guessing it's checking the context and not finding a reference to
the extension. So we'll add this and, hopefully, it should solve things.