This feels really fragile and like it could be designed better but things work, for now.
I still need to handle if someone clicks the button a second time…but
one bit at a time.

Maybe I'll move on to cleaning up the joins, in the backend, or
something.

And then, now that we have properly signing HTTP calls down, we can
try actually creating posts and the like; or, at least, responding to posts.